# internal/httpapi/gen/api.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for EventActorType.
const (
	EventActorTypeAgent  EventActorType = "agent"
	EventActorTypeSystem EventActorType = "system"
	EventActorTypeUser   EventActorType = "user"
)

// Valid indicates whether the value is a known member of the EventActorType enum.
func (e EventActorType) Valid() bool {
	switch e {
	case EventActorTypeAgent:
		return true
	case EventActorTypeSystem:
		return true
	case EventActorTypeUser:
		return true
	default:
		return false
	}
}

// Defines values for HealthDatabase.
const (
	HealthDatabaseOk          HealthDatabase = "ok"
	HealthDatabaseUnreachable HealthDatabase = "unreachable"
)

// Valid indicates whether the value is a known member of the HealthDatabase enum.
func (e HealthDatabase) Valid() bool {
	switch e {
	case HealthDatabaseOk:
		return true
	case HealthDatabaseUnreachable:
		return true
	default:
		return false
	}
}

// Defines values for HealthStatus.
const (
	HealthStatusDegraded HealthStatus = "degraded"
	HealthStatusOk       HealthStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case HealthStatusDegraded:
		return true
	case HealthStatusOk:
		return true
	default:
		return false
	}
}

// Defines values for Role.
const (
	Admin  Role = "admin"
	Viewer Role = "viewer"
)

// Valid indicates whether the value is a known member of the Role enum.
func (e Role) Valid() bool {
	switch e {
	case Admin:
		return true
	case Viewer:
		return true
	default:
		return false
	}
}

// ChangePasswordRequest defines model for ChangePasswordRequest.
type ChangePasswordRequest struct {
	CurrentPassword string `json:"current_password"`
	NewPassword     string `json:"new_password"`
}

// Error defines model for Error.
type Error struct {
	// Code Stabiele foutcode, bijvoorbeeld invalid_credentials of totp_required
	Code    string `json:"code"`
	Message string `json:"message"`
}

// Event defines model for Event.
type Event struct {
	Action      string                 `json:"action"`
	ActorId     string                 `json:"actor_id"`
	ActorType   EventActorType         `json:"actor_type"`
	ClusterId   *openapi_types.UUID    `json:"cluster_id,omitempty"`
	Id          int64                  `json:"id"`
	Payload     map[string]interface{} `json:"payload"`
	SubjectId   string                 `json:"subject_id"`
	SubjectType string                 `json:"subject_type"`
	Ts          time.Time              `json:"ts"`
}

// EventActorType defines model for Event.ActorType.
type EventActorType string

// Health defines model for Health.
type Health struct {
	Database HealthDatabase `json:"database"`
	Status   HealthStatus   `json:"status"`
	Version  string         `json:"version"`
}

// HealthDatabase defines model for Health.Database.
type HealthDatabase string

// HealthStatus defines model for Health.Status.
type HealthStatus string

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`

	// TotpCode Zescijferige code
	TotpCode *string `json:"totp_code,omitempty"`
	Username string  `json:"username"`
}

// Me defines model for Me.
type Me struct {
	CsrfToken        string    `json:"csrf_token"`
	SessionExpiresAt time.Time `json:"session_expires_at"`
	User             User      `json:"user"`
}

// Role defines model for Role.
type Role string

// TotpCodeRequest defines model for TotpCodeRequest.
type TotpCodeRequest struct {
	Code string `json:"code"`
}

// TotpDisableRequest defines model for TotpDisableRequest.
type TotpDisableRequest struct {
	Code     string `json:"code"`
	Password string `json:"password"`
}

// TotpSetup defines model for TotpSetup.
type TotpSetup struct {
	OtpauthUrl string `json:"otpauth_url"`
	Secret     string `json:"secret"`
}

// User defines model for User.
type User struct {
	CreatedAt   time.Time          `json:"created_at"`
	Id          openapi_types.UUID `json:"id"`
	Role        Role               `json:"role"`
	TotpEnabled bool               `json:"totp_enabled"`
	Username    string             `json:"username"`
}

// ListEventsParams defines parameters for ListEvents.
type ListEventsParams struct {
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// ChangePasswordJSONRequestBody defines body for ChangePassword for application/json ContentType.
type ChangePasswordJSONRequestBody = ChangePasswordRequest

// DisableTotpJSONRequestBody defines body for DisableTotp for application/json ContentType.
type DisableTotpJSONRequestBody = TotpDisableRequest

// EnableTotpJSONRequestBody defines body for EnableTotp for application/json ContentType.
type EnableTotpJSONRequestBody = TotpCodeRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Login Inloggen met wachtwoord en, indien ingesteld, TOTP-code
	// (POST /auth/login)
	Login(w http.ResponseWriter, r *http.Request)
	// Logout Uitloggen
	// (POST /auth/logout)
	Logout(w http.ResponseWriter, r *http.Request)
	// GetMe De ingelogde gebruiker en de CSRF-token van de sessie
	// (GET /auth/me)
	GetMe(w http.ResponseWriter, r *http.Request)
	// ChangePassword Wachtwoord wijzigen; andere sessies worden beëindigd
	// (POST /auth/password)
	ChangePassword(w http.ResponseWriter, r *http.Request)
	// DisableTotp TOTP uitzetten met wachtwoord en een geldige code
	// (POST /auth/totp/disable)
	DisableTotp(w http.ResponseWriter, r *http.Request)
	// EnableTotp TOTP activeren met een code uit de authenticator-app
	// (POST /auth/totp/enable)
	EnableTotp(w http.ResponseWriter, r *http.Request)
	// BeginTotpSetup Nieuw TOTP-geheim aanmaken (nog niet actief)
	// (POST /auth/totp/setup)
	BeginTotpSetup(w http.ResponseWriter, r *http.Request)
	// ListEvents Recente events, nieuwste eerst (alleen admin)
	// (GET /events)
	ListEvents(w http.ResponseWriter, r *http.Request, params ListEventsParams)
	// GetHealth Gezondheid van de server en de database
	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Login Inloggen met wachtwoord en, indien ingesteld, TOTP-code
// (POST /auth/login)
func (_ Unimplemented) Login(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Logout Uitloggen
// (POST /auth/logout)
func (_ Unimplemented) Logout(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetMe De ingelogde gebruiker en de CSRF-token van de sessie
// (GET /auth/me)
func (_ Unimplemented) GetMe(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ChangePassword Wachtwoord wijzigen; andere sessies worden beëindigd
// (POST /auth/password)
func (_ Unimplemented) ChangePassword(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// DisableTotp TOTP uitzetten met wachtwoord en een geldige code
// (POST /auth/totp/disable)
func (_ Unimplemented) DisableTotp(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// EnableTotp TOTP activeren met een code uit de authenticator-app
// (POST /auth/totp/enable)
func (_ Unimplemented) EnableTotp(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// BeginTotpSetup Nieuw TOTP-geheim aanmaken (nog niet actief)
// (POST /auth/totp/setup)
func (_ Unimplemented) BeginTotpSetup(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ListEvents Recente events, nieuwste eerst (alleen admin)
// (GET /events)
func (_ Unimplemented) ListEvents(w http.ResponseWriter, r *http.Request, params ListEventsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetHealth Gezondheid van de server en de database
// (GET /health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Login(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetMe operation middleware
func (siw *ServerInterfaceWrapper) GetMe(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ChangePassword operation middleware
func (siw *ServerInterfaceWrapper) ChangePassword(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ChangePassword(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DisableTotp operation middleware
func (siw *ServerInterfaceWrapper) DisableTotp(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DisableTotp(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// EnableTotp operation middleware
func (siw *ServerInterfaceWrapper) EnableTotp(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.EnableTotp(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BeginTotpSetup operation middleware
func (siw *ServerInterfaceWrapper) BeginTotpSetup(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BeginTotpSetup(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListEvents operation middleware
func (siw *ServerInterfaceWrapper) ListEvents(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ListEventsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListEvents(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/login", wrapper.Login)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/logout", wrapper.Logout)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/auth/me", wrapper.GetMe)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/password", wrapper.ChangePassword)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/totp/setup", wrapper.BeginTotpSetup)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/totp/enable", wrapper.EnableTotp)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/totp/disable", wrapper.DisableTotp)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/events", wrapper.ListEvents)
	})

	return r
}