# internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package store

import (
	"database/sql/driver"
	"fmt"
	"time"

	"github.com/google/uuid"
)

type ActorType string

const (
	ActorTypeUser   ActorType = "user"
	ActorTypeAgent  ActorType = "agent"
	ActorTypeSystem ActorType = "system"
)

func (e *ActorType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = ActorType(s)
	case string:
		*e = ActorType(s)
	default:
		return fmt.Errorf("unsupported scan type for ActorType: %T", src)
	}
	return nil
}

type NullActorType struct {
	ActorType ActorType
	Valid     bool // Valid is true if ActorType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullActorType) Scan(value interface{}) error {
	if value == nil {
		ns.ActorType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.ActorType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullActorType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.ActorType), nil
}

type UserRole string

const (
	UserRoleAdmin  UserRole = "admin"
	UserRoleViewer UserRole = "viewer"
)

func (e *UserRole) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = UserRole(s)
	case string:
		*e = UserRole(s)
	default:
		return fmt.Errorf("unsupported scan type for UserRole: %T", src)
	}
	return nil
}

type NullUserRole struct {
	UserRole UserRole
	Valid    bool // Valid is true if UserRole is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullUserRole) Scan(value interface{}) error {
	if value == nil {
		ns.UserRole, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.UserRole.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullUserRole) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.UserRole), nil
}

type Event struct {
	ID          int64
	Ts          time.Time
	ActorType   ActorType
	ActorID     string
	SubjectType string
	SubjectID   string
	ClusterID   *uuid.UUID
	Action      string
	Payload     []byte
}

type Session struct {
	ID         []byte
	UserID     uuid.UUID
	CsrfToken  string
	Ip         string
	UserAgent  string
	CreatedAt  time.Time
	LastSeenAt time.Time
	ExpiresAt  time.Time
}

type User struct {
	ID            uuid.UUID
	Username      string
	PasswordHash  string
	Role          UserRole
	TotpSecret    *string
	TotpEnabledAt *time.Time
	TotpLastStep  *int64
	DisabledAt    *time.Time
	CreatedAt     time.Time
	UpdatedAt     time.Time
}